requirements.txt
python-whois
pyyaml
requests
tqdm
dnspython